requirements.txt
flask
flask-cors
deepforest
torch
Pillow
tensorflow
keras
opencv-python-headless
requests
